+++ Basil/include/VertexBuffer.h
@@ -6,7 +6,6 @@
 #include <cstdint>
 
 #include "BufferLayout.h"
-#include "Renderer.h"
 
 namespace Basil
 {