+++ Basil/src/VertexBuffer.cpp
@@ -1,6 +1,7 @@
 #include "pch.h"
 #include "VertexBuffer.h"
 #include "OpenGLVertexBuffer.h"
+#include "Renderer.h"
 
 namespace Basil
 {
@@ -9,10 +10,10 @@ namespace Basil
 	{
 		switch (Renderer::getAPI())
 		{
-			case RendererAPI::None:
+			case RendererAPI::API::None:
 				ASSERT(false, "No renderer API!");
 				return nullptr;
-			case RendererAPI::OpenGL:
+			case RendererAPI::API::OpenGL:
 				return new OpenGLVertexBuffer(vertices);
 		}